# Use control point in preference to intersections