… internal/cmprefimpl/cmpgeos/checks.go