## Description

    Insert ctrl points before intersection points when renoding
    
    This is to give preference to control points over intersection points in
    outputs (in the case where the points would be aliased).

## Check List

Have you:

- Added unit tests? Yes

- Add cmprefimpl tests? (if appropriate?) N/A

## Related Issue

- https://github.com/peterstace/simplefeatures/issues/283

## Benchmark Results

```
COMPARISON
name                    old time/op    new time/op    delta
Intersection/n=10-2        221µs ±25%     220µs ± 8%    ~     (p=0.650 n=15+13)
Intersection/n=100-2      1.86ms ±12%    1.88ms ± 7%    ~     (p=0.701 n=14+14)
Intersection/n=1000-2     20.8ms ± 9%    20.2ms ± 6%    ~     (p=0.106 n=14+12)
Intersection/n=10000-2     245ms ± 4%     240ms ± 9%  -2.06%  (p=0.011 n=12+14)

name                    old alloc/op   new alloc/op   delta
Intersection/n=10-2       57.3kB ± 0%    57.4kB ± 0%  +0.33%  (p=0.000 n=15+15)
Intersection/n=100-2       504kB ± 0%     505kB ± 0%  +0.03%  (p=0.029 n=15+15)
Intersection/n=1000-2     6.18MB ± 0%    6.18MB ± 0%    ~     (p=0.798 n=15+15)
Intersection/n=10000-2    60.1MB ± 0%    60.1MB ± 0%    ~     (p=0.412 n=15+15)

name                    old allocs/op  new allocs/op  delta
Intersection/n=10-2          747 ± 0%       753 ± 0%  +0.79%  (p=0.000 n=15+15)
Intersection/n=100-2       5.61k ± 0%     5.62k ± 0%  +0.10%  (p=0.000 n=15+15)
Intersection/n=1000-2      53.7k ± 0%     53.7k ± 0%    ~     (p=0.321 n=13+15)
Intersection/n=10000-2      534k ± 0%      534k ± 0%    ~     (p=0.911 n=15+15)
```